Reworked parts of the Routines
Routines execute when the time for their execution was reached once. There's also a mechanism to allow routines to stack and wait for their previous counterpart to finish. This makes it possible to have an NPC start a routine only a few minutes apart, but still complete them in the right order - without skipping a single one.
(Godfrey can now roam the entire map)